Fix: string_property_editor_screen: delete character shortens string
Deleting a character causes the string to shift left. However, it
shouldn't shift the terminating null as the string is supposed to be
entirely filled with spaces when it is used by the editor.

This confused the clean-up code, leaving spaces in the edited string.

+++ include/display/string_property_editor.hpp
@@ -30,7 +30,7 @@ public:
 
 	void set_property(const __FlashStringHelper *prompt,
 			  char *property,
-			  size_t property_size) noexcept;
+			  uint8_t property_size) noexcept;
 
 	void focused() noexcept override;
 
@@ -63,7 +63,7 @@ private:
 	const __FlashStringHelper *_prompt;
 	struct {
 		char *value;
-		size_t size;
+		uint8_t size;
 	} _property;
 
 	uint8_t _focused_character;
